временный

+++ service/model/ticket-chat-id.go
@@ -2,5 +2,5 @@ package model
 
 type TicketChatId struct {
 	TicketId int32
-	ChatId   int32
+	ChatId   int64
 }
